--- bench/bench.go
@@ -1,33 +1,0 @@
-/*
- * Copyright 2019 Dgraph Labs, Inc. and Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package bench
-
-// Cache needs to be fulfilled by the cache implementations in order for the
-// benchmarks to run properly.
-type Cache interface {
-	Get(string) interface{}
-	Set(string, interface{})
-	Del(string)
-	Bench() *Stats
-}
-
-// Stats holds hit/miss information after a round of benchmark iterations has
-// been ran.
-type Stats struct {
-	Reqs uint64
-	Hits uint64
-}
