--- bench/cache_basemutex.go
@@ -1,56 +1,0 @@
-/*
- * Copyright 2019 Dgraph Labs, Inc. and Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package bench
-
-import (
-	"sync"
-
-	"github.com/golang/groupcache/lru"
-)
-
-type BenchBaseMutex struct {
-	sync.Mutex
-	cache *lru.Cache
-}
-
-func NewBenchBaseMutex(capacity int) *BenchBaseMutex {
-	return &BenchBaseMutex{
-		cache: lru.New(capacity),
-	}
-}
-
-func (c *BenchBaseMutex) Get(key string) interface{} {
-	c.Lock()
-	defer c.Unlock()
-	value, _ := c.cache.Get(key)
-	// value found
-	return value
-}
-
-func (c *BenchBaseMutex) Set(key string, value interface{}) {
-	c.Lock()
-	defer c.Unlock()
-	c.cache.Add(key, value)
-}
-
-func (c *BenchBaseMutex) Del(key string) {
-	c.cache.Remove(key)
-}
-
-func (c *BenchBaseMutex) Bench() *Stats {
-	return nil
-}
